scripts/requirements.txt
fastapi
uvicorn
psycopg[binary]>=3.0
python-dotenv
requests
